Add crawler as a container
