+++ compose.yaml
@@ -1,4 +1,31 @@
 services:
+  dashboard:
+    image: webmap-dashboard
+    environment:
+      - NEO4j_URI=neo4j://database
+      - NEO4j_USERNAME=neo4j
+      - NEO4j_PASSWORD=password
+    ports:
+      - "8000:8000"
+    networks:
+      - webmap-network
+    depends_on:
+      - database
+    restart: on-failure
+
+  crawler:
+    image: webmap-crawler
+    environment:
+      - NEO4j_URI=neo4j://database
+      - NEO4j_USERNAME=neo4j
+      - NEO4j_PASSWORD=password
+    networks:
+      - webmap-network
+    depends_on:
+      - database
+      - selenium
+    restart: on-failure
+
   database:
     image: neo4j:latest
     environment:
@@ -10,22 +37,22 @@ services:
       - neo4j_data:/data
     networks:
       - webmap-network
-    restart: always
+    restart: unless-stopped
 
-  dashboard:
-    image: webmap-dashboard
-    environment:
-      - DATABASE_URL="NA"
-      - DATABASE_PORT="NA"
+  selenium:
+    image: selenium/standalone-chrome
+    volumes:
+      - /dev/shm:/dev/shm 
+      - chrome_profiles:/tmp/chrome
     ports:
-      - "8002:8000"
+      - 4444:4444
     networks:
       - webmap-network
-    restart: "no"
-
+    restart: unless-stopped
 
 volumes:
   neo4j_data:
+  chrome_profiles:
 
 networks:
   webmap-network:
